requirements.txt
streamlit
psycopg2-binary
sqlalchemy
pandas
requests
plotly
matplotlib
